Extract business_model_canvas/ bounded context
Move BmcProjectPresenter into business_model_canvas/presenter.py.
bin/cli/infrastructure/bmc_presenter.py becomes a re-export shim.

Part of #47.
